Add project
Read in list of domains or urls and then test the urls to see if the
website exists.  If it exists then then it will return a http response
(200 if the website is working).  If the website does not exist then it
will throw a WebExecption error.  Take these results and write them to a
file.